made the sending manager and lever separate and implemented in scene
