+++ Assets/Scripts/SC_SendingManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3b28441c2686f734e8df0e80cb46d7b5
